Update patch set 1
Patch Set 1:

i agree that we don't need the multimode support, and your suggested reasoning for how we got in this situation sounds likely.

the test failures from the bug in the commit message were:

junit.framework.AssertionFailedError: Failed to round properly - up to odd expected:<3.0> but was:<2.625>
       at org.apache.harmony.luni.tests.java.lang.StrictMathTest.test_rintD(StrictMathTest.java:899)

junit.framework.AssertionFailedError: Should return 0.4633708685409921 expected:<0.4633708685409921> but was:<0.46337086854099213>
       at org.apache.harmony.luni.tests.java.lang.StrictMathTest.test_log1p_D(StrictMathTest.java:573)

so that should be easy to test. i'll have a look...

Patch-set: 1